requirements.txt
fastapi==0.116.1
uvicorn==0.35.0
openai==1.98.0
google-genai==1.28.0
requests==2.32.4
beautifulsoup4==4.13.4
playwright==1.54.0
pandas==2.3.1
numpy==2.3.2
duckdb==1.3.2
matplotlib==3.10.5
python-dotenv==1.1.1
PyPDF2==3.0.1
python-magic==0.4.27
Pillow==11.3.0
python-multipart==0.0.20
pytest==8.3.2
pytest-mock==3.14.0
pytest-asyncio==0.23.8

scikit-learn==1.7.1
scipy==1.16.1
lxml==6.0.0
